pyproject.toml
[project]
name = "coop-ie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "playwright>=1.49.0",
    "pydantic>=2.10.4",
    "pydantic-ai[logfire]>=0.0.12",
    "python-dotenv>=1.0.1",
]
